Allow model-input from user for gridding
Automatic gridding happens at the first call to
`Simulation3DEMG3D.emg3d_sim()`. Until now, a dummy model with 1 S/m was
provided (same grid was used for computation as for inversion).

For automatic gridding (`gridding!='same'`), this has to be a realistic
model. The user can now provide a model with the `simulation_opts`.
Usually this will be the reference or the starting model.