Avoid using getattr with fixed constant strings
Fix `flake8` `B008` warnings. Refactor the way fixed attributes are
obtained: instead of using `getattr` with fixed strings, use the
standard `object.attribute` syntax that is more efficient and readable.
Use `hasattr` for checking if the wanted attribute exists instead of
using `getattr` and returning None if not: improves readability.