Fallback to original try-except for meshes
Instead of checking if `meshes` is `Iterable`, we decided to check if we
can actually unpack it in two elements and throw a `TypeError` while
catching a `TypeError` on the `except`.
Remove the test for the length of `meshes`: if it has other than two
elements the first `try` will fail.