Run flake8-permissive with ignore rather than select
Instead of selecting a few rules that should be checked it might be
better to flag the ones we really want to ignore. This allows
`flake-permissive` to check if any other rule that is not included in
the list is being violated. Had to override the default ignored rules by
flake8, so the default ones are also included in the list.