Generalize __add__  for any ComboObjectiveFunction
Rewrite the `__add__` method of `BaseObjectiveFunction` so by default
it performs the same task for any `ComboObjectiveFunction` or any of its
child classes. Add a new `unpack_on_add` attribute to
`ComboObjectiveFunction` to decide weather to unpack its objective
functions after an addition, or if it should be added as a whole. Set
`unpack_on_add=True` by default on `ComboObjectiveFunction`, but set it
to `False` for regularization classes, like `WeightedLeastSquares` and
`PGI`. Add tests for the new changes.