Remove indActive from WeightedLeastSquares kwargs
Remove `indActive` from `WeightedLeastSquares` kwargs and assign it to
`active_cells`. Explicitly set `self.active_cells` after calling
parent's `__init__`. This simplifies the process of setting
active_cells: previously the `indActive` argument was being passed to
every parent up to `BaseObjectiveFunction`, where the attr was set. This
process was a little bit obscure: setting the `active_cells` attribute
needs the `objcfts` attribute already defined, so the order in which
they are defined is important, but it's decided at parent's levels, and
not in `WeightedLeastSquares`.