Move flake8 ignores to .flake8
Moves the flake rules we are currently ignoring to the .flake8
configuration file. Rename the Makefile targets: `make flake` runs
flake8 with ignores set in `.flake8`, while `make flake-full` runs
flake8 with the full set of rules (without ignoring the ones we aren't
following at the moment). Add a `help` target to the `Makefile` that
prints out description of the available targets.