Allow to use random seed in make_synthetic_data (#1286)
Add a new random_seed argument to the make_synthetic_data method.
Switch to the Numpy's random number generator object and pass the
random_seed argument to np.random.default_rng, allowing users to
specify a random seed for the synthetic noise. Set random seeds in some
tests that were failing due to the new random state that the usage of
make_synthetic_data with random_seed created in the test suite.