Gravity simulation with Choclo as engine (#1285)
Add the option to use [Choclo](https://www.fatiando.org/choclo) as the
engine for running the gravity simulations, considerably reducing the
computation time of the forward simulations and the creation of the
sensitivity matrix. Add Choclo as one of the optional dependencies.
Update and extend tests for the gravity simulation using pytest. By
default, the simulation will use `geoana` as the engine. Users can
alternatively choose `choclo` instead if it's installed.