Remove deprecated regularization classes (#1365)
Raise errors when constructing deprecated regularization classes. Add
test that checks if those errors are being raised. Replace the removed
regularization classes that were still being used in a few pieces of the
code and in the tests. Use the new classes instead. Add removed classes
to the `IGNORE_ME` list in `test_regularization.py` so other tests don't
fail.