Enforce regularization `weights` as dictionaries (#1344)
Only allow passing `weights` to regularizations as `None` or
dictionaries. Raise error if `weights` are a different type. Remove
lines that were assigning weights if the passed argument was an array.
Update examples, tests and tutorials to follow the new interface. Update
how the `PGI` constructor passes weights to the underlying
`WeightedLeastSquares` objects.

---------

Co-authored-by: Santiago Soler <santisoler@fastmail.com>
Co-authored-by: yinghu <huying.em@gmail.com>