Dask MetaSim (#1199)
Adds the `dask` version of the `MetaSimulation`.

Essentially it never explicitly sends around fields objects, and it
avoids sending around simulations (after they have gone through the
setter). Every operation that involves one of the internal simulations
is explicitly done on the worker that owns that simulation. The `fields`
returned by this class are actually `Future`s that live on the worker
that owns the respective simulation.

Most items that live on this class are `Future`s of the respective
item in the `MetaSimulation` class ( e.g. `sim.mappings` are a list
of `Future` maps instead of a `list` of maps). As a side effect, it is
actually possible to initialize the class with lists of Futures. This is
useful if you want to avoid creating everything on the main thread and
then serializing and sending the result out to each worker, instead you
can create a function which returns an object and send that function to
each worker.