Bump Black version to 24.3.0 (#1403)
Update black version to 24.3.0 in `environment_test.yml`, requirements
files and the configuration file for `pre-commit`. This update was
triggered after a moderate vulnerability to Regular Expression Denial of
Service (ReDoS). The vulnerability was patched in Black 24.3.0. Run new
version of black on code base and docs. Replace lambda definition for
function in test due to flake8 error after formatted by the new version
of Black.