Add seed argument to objective fun's derivative tests
Add a new `seed` argument to the `test()` method of objective functions
to control their random state. Use Numpy's random number generator for
managing the randoms state and the generation of random numbers. Minor
improvements to the implementation of the test methods. Update tests
that make use of these methods, and make them to use a seed in every
case.