Stop recommending mamba for installing simpeg
Since latest versions of `conda` make use of `libmamba` that
achieves the same performance as `mamba` while installing packages and
creating environments, we can stop recommending `mamba` to our users
as a better alternative for `conda`. Update the admonitions in the
instructions to install simpeg and to set up the environment for
contributors. Mention the minimum version of `conda` that uses
`libmamba` by default. Replace recommendations for Mambaforge to
Miniforge.