Simplify interface of UniformBackgroundField (#1421)
Remove the `**kwargs` from the constructor of `UniformBackgroundField`.
Make `receiver_list`, `amplitude`, `inclination` and `declination` as
required arguments without any default value. Include type hints in the
signature of the constructor. Add check for type of receivers to the
constructor of `UniformBackgroundField`. Add test that checks the value
of `UniformBackgroundField.b0`, passing `receiver_list` as an actual
list of point receivers and as `None`. Add test that checks if the error
is raised after passing invalid receiver type. Minor improvements to the
docstring of the class.