Fix bug on arguments of beta estimator directives (#1460)
Fix bug in the constructor of beta estimator directive classes: the
`seed` argument was being passed as a positional argument to the
parent's constructor, which would assign it to the `n_pw_iter` attribute
instead to the `seed`, leading to unwanted behaviours. Pass the
arguments to the constructor through keywords to make it more explicit
and less prone to errors. Add tests that would fail without this bugfix.
Remove the `n_pw_iter` and the `method` arguments from the constructor
of `BaseBetaEstimator`.