Improve interface for DC Dipole source (#1393)
Remove `**kwargs` from the constructor of the class, add `current` as
one of the arguments. Improve the logic and error messages for the
`location`, `location_a` and `location_b` arguments. Add tests for these
new checks. Update docstring of the constructor to reflect latest
changes. Fix the expected shape of the `location_a` and `location_b`
arguments.