Magnetic simulation with Choclo as engine (#1321)
Add new implementation of the magnetic simulation using the integral
formulation based on [Choclo](https://www.fatiando.org/choclo) and
Numba. Allow users to run the simulation with Choclo as engine through a
new `engine` argument of the constructor of the simulation class. Add
`numba_parallel` argument to control if the simulation should be run in
parallel or in a single thread. Move common methods and properties of
the gravity and magnetic simulation to the base class for potential
field simulations. Add new `potential_fields/_numba_utils.py` file with
functions that are needed both in the magnetic and gravity Numba
implementations. Refactor and extend tests for the magnetic simulation,
including tests that assert the accuracy of the new implementation
against analytic solutions. Add new tests for the base class for
potential field simulations. Update examples using Choclo as engine,
include admonitions instructing how to use it.

---------

Co-authored-by: Jacob Edman <jacob.edman@koboldmetals.com>
Co-authored-by: Joseph Capriotti <josephrcapriotti@gmail.com>