Add version switcher to Sphinx docs (#1428)
Add a version switcher to the navbar of the docs, using the built in one
in PyData Sphinx theme. Point to the experimental `versions.json` file
located in `doctest.simpeg.xyz` that should serve the files in
[simpeg/simpeg-doctest](https://github.com/simpeg/simpeg-doctest). Add
experimental staged to Azure Pipelines that pushes docs to the
`simpeg-doctest` repository. One is triggered nightly and deploys the
latest docs in `main` to the `dev` branch in `simpeg-doctest` and
updates the `dev` submodule in `gh-pages`. The other one is triggered
after a release and pushes the new version of the docs to `gh-pages`
branch in `simpeg-doctest` onto a new folder, while also updating the
`latest` link.