Fix links to source code in documentation pages (#1444)
Update the `linkcode_resolve` function in Sphinx configuration file so
the links to the source code of each function, class or method points to
the actual code corresponding to that version, and not always to the
code in `main`. Update some Azure configurations: checkout simpeg repo
without shallow depth and fetching tags. Clean up the working directory
before installing SimPEG so the version doesn't have an extra hash.

---------

Co-authored-by: Joseph Capriotti <josephrcapriotti@gmail.com>