Fix script for new deployment of docs (#1478)
Fix issues in the new (experimental) scripts for deployment of docs to
`simpeg-doctest`. Fetch the `gh-pages` before trying to checkout it.
This is needed because we clone the repo with `--depth 1`. When passing
this option, git appends the `--single-branch` option meaning the `dev`
branch is the only branch that is being cloned. Improve how we clone the
`simpeg-doctest` repo: clone quietly to reduce output lines, and specify
the branch while deploying release docs (in case the default branch in
the repo gets changed).