Add `random_seed` argument to objective fun's derivative tests (#1448)
Add a new `random_seed` argument to the `test()` method of objective
functions to control their random state. Use Numpy's random number
generator for managing the random state and the generation of random
numbers. Minor improvements to the implementation of the test methods.
Update tests that make use of these methods, and make them to use a seed
in every case.

Part of the solution to #1289