Add `random_seed` to the `test` method of maps (#1465)
Add `random_seed` argument to `IdentityMap.test`. Add a new
`random_seed` argument to the `test` method of `IdentityMap` to control
the random state for creating the `m` array if it's not provided. Update
tests to use the `random_seed`. If the `m` array is passed while running
the `test()` method, specify the keyword for the argument to avoid
relying only on its position.