Fix extraction of files in some examples
Fix how the tar.gz archives that get downloaded are extracted in a few
different examples. The previous code was getting the folder where files
are extracted by splitting the full path of the downloaded .tar.gz
archive using a `"."`. This causes issues when any parent folder
contains a dot: the directory that is returned is not the correct one.