Ignore `survey_type` argument in DC and SIP surveys (#1458)
Ignore the `survey_type` argument in the dc and sip survey constructors
and raise warning. Raise warning when trying to access or set the
`survey_type` properties. The `survey_type` argument is not being used
anymore and was a residue from an older implementation of the survey
classes. Keeping it as part of the class creates confusion among users
on how to properly use it. Update tests and examples that were still
using that argument. Add tests to check if warnings are being raised
when trying to access the `survey_type` property.

Fixes #1432