Update deployment of docs to simpeg-docs (#1490)
Remove the `old-docs.yml` Azure pipeline. Update the new pipelines to
deploy docs so they push the changes to the `gh-pages` and `dev`
branches in the `simpeg-docs` repo. Update the url for the
`versions.json`. Move scripts that deploy the docs to actual bash
scripts. Apply some improvements to the scripts and solve a few issues
in them that weren't working as expected.