1454 suggestions (#1507)
#### Summary
Suggestions for #1454

#### What does this implement/fix?
Restructures the classes and fixes several errors.

#### Additional information

There's a bit here that I am suggesting to change.
First some errors that were in the original PR:
* Removes unused `return_complex` argument that was defined on the
constructors in favor of passing `complex` to the component argument
(Throws an error `eval_deriv` if `component=='complex'`).
* Restructures the inheritance using a new base class:
    * New base `BaseNaturalSourceRx`
* handles setting location pairs and ensuring they're matching pairs
(both array_like, with the same shape).
* significantly simplifies the `getP` method (instead of referring to
each location array by a name, it passes the requested index).
        * Ensures `nD` is properly described.
* Classes no longer inherit from the `ApparentConductivity` class that
is used to simulate the mobile MT data.
* Old classes inherit from the new classes so that our internal code can
expect everything to work like the new classes, but people will "mostly"
be able to quickly change them.
* Instead of making the new classes and the old classes function
identically (as required previously).
* Uses the `assert_isadjoint` test function from discretize (which looks
to have exposed something incorrect for the adjoint in the `Tipper`
class).

####  `locations`

I am suggesting one breaking change though, and I think it's worthwhile,
and that is that the new receivers `locations` property will now be a
`tuple` of `ndarray` (one for each field type). Looking at how the
previous receivers handled setting these location pairs makes me think
that there were just too many issues with how it was set previously (and
even that was inconsistent with itself).

For example, If you did the following (on `main`):

```python
>>> locations_e = [[0, 0, 0]]
>>> locations_h = [[1, 1, 1]]
>>> rx = PointNaturalSource(locations=[locations_e, locations_h])
>>> rx.locations
array([[0, 0, 0]])
>>> rx.locations_e
array([[0, 0, 0]])
>>> rx.locations_h
array([[1, 1, 1]])
```

if you did instead this:
```python
>>> locations_e = [[0, 0, 0]]
>>> locations_h = [[1, 1, 1]]
>>> rx = PointNaturalSource(locations_e=locations_e, locations_h=locations_h)
>>> rx.locations
array([[0, 0, 0, 1, 1, 1]])
>>> rx.locations_e
array([[0, 0, 0]])
>>> rx.locations_h
array([[1, 1, 1]])
```
but they would both create the same data. So `locations` as a property
was already kinda broken, and this fixes it to be consistent.