Use `packaging` to check scipy version
Make use of `packaging.version.parse` to check if the available version
of scipy is greater or equal than 1.12. Add `packaging` as a required
dependency. Minor improvements of code style: use uppercase for the
global variable, define key for the tolerance as a variable and define
the `inp` dictionary only once.