Replace `indActive` and `actInd` for `active_cells` in maps (#1534)
Deprecate the `indActive`, `actInd` arguments and properties from maps
and replace them for `active_cells`. Deprecate the `valInactive`
argument and property from `InjectActiveCells` and replace it for
`value_inactive`. Add tests for the deprecations. Avoid a try and except
pattern in the setter for `value_inactive`: we don't want to ignore any
error that might happen in those lines. Update usage of `actInd`,
`indActive` and `valInactive` in tutorials and examples.

Part of #1121