Default solver (#1511)
#### Summary
Adds methodology to explicitly set a default solver for simpeg to use,
and explicitly lists a priority to try to import them.

#### What does this implement/fix?
There are too many explicit uses of `Pardiso` in the simpeg tests and
examples making them unrunnable on systems that do not implement
`Pardiso`.

It is also unclear to the user which solver is chosen when using
default, thus if the user implicitly uses a default solver in a
simulation, or asks for the default solver, a `DefaultSolverWarning` is
issued saying explicitly what solver is used. This is a new Warning
class that the end user can choose to silence if needed.

This also removes the re-implemented Solver wrappers and instead rely on
`pymatsolver`'s versions of those wrappers.