Numpy2.0 and discretize 0.11.0 updates (#1558)
#### Summary
Necessary updates for `numpy2.0` and `discretize` 0.11.0

#### PR Checklist
* [x] If this is a work in progress PR, set as a Draft PR
* [x] Linted my code according to the [style
guides](https://docs.simpeg.xyz/latest/content/getting_started/contributing/code-style.html).
* [x] Added
[tests](https://docs.simpeg.xyz/latest/content/getting_started/contributing/testing.html)
to verify changes to the code.
* [x] Added necessary documentation to any new functions/classes
following the
expect
[style](https://docs.simpeg.xyz/latest/content/getting_started/contributing/documentation.html).
* [x] Marked as ready for review (if this is was a draft PR), and
converted
      to a Pull Request
* [x] Tagged ``@simpeg/simpeg-developers`` when ready for review.

#### What does this implement/fix?
Related to Numpy 2.0:
* Simplifies `unique_rows` function to more directly use numpy's
`unique` function.
Related to discretize 0.11:
* Uses public `get_containing_cells` function of `TreeMesh` instead of
private method.
* Adds `random_seed` arguments to `check_derivative` and
`assert_isadjoint` functions now that it is supported.

#### Additional information
We were generally in a good position for numpy 2.0! The only breaking
point from discretize 0.11 was due to using a private method of a
TreeMesh, which is again a pretty good place to be. It's expected that
the `check_derivative` tests need to updated given the new random
generator usage.