Only warn about default solver when set in simulations
Add a `warn=False` argument to the `get_default_solver()` function. Only
when `warn=True`, the warning about the setting of the default solver is
raised. This makes the function not to raise a warning when called with
the default arguments. When default solvers are being used within
simulations, call the function with `warn=True` to let users know about
the solver being automatically selected. Update tests accordingly.