Remove empymod dependency (#1571)
#### Summary
`empymod` is a pretty heavy dependency for the functionality we are
using from it. This removes the dependency and instead uses `libdlf` to
get the filters, and a streamlined `get_splined_dlf_points` for our
internal use.

#### PR Checklist
* [x] If this is a work in progress PR, set as a Draft PR
* [x] Linted my code according to the [style
guides](https://docs.simpeg.xyz/latest/content/getting_started/contributing/code-style.html).
* [x] Added
[tests](https://docs.simpeg.xyz/latest/content/getting_started/contributing/testing.html)
to verify changes to the code.
* [x] Added necessary documentation to any new functions/classes
following the
expect
[style](https://docs.simpeg.xyz/latest/content/getting_started/contributing/documentation.html).
* [x] Marked as ready for review (if this is was a draft PR), and
converted
      to a Pull Request
* [x] Tagged ``@simpeg/simpeg-developers`` when ready for review.

#### Additional information
I say `empymod` is a heavy dependency because it in turn relies on
`numba`.