Bugfix for TDEM magnetic dipole sources  (#1572)
#### Summary
Bugfix so that the initial fields produced for `db/dt`, `dh/dt`, from
the B field or H field simulations are zero. There was previously a bug
where we set the electric source term to `Zero()` when infact it should
have been evaluated. This would only effect `db/dt` (or `dh/dt`) that
were measured before the first time-step, but it also impacted
visualizations of the fields.

To catch this issue, I have extended our cross-check tests for tdem
simulation to also include an evaluation of the data at t=0. Running
this test on the current main version of SimPEG will fail.