Rename delete on model update (#1589)
#### Summary
Privatizes the `deleteOnModelUpdate` property, and unify with the
`clean_on_model_update`.

#### Reference issue
Working towards #1429

#### What does this implement/fix?
The `deleteOnModelUpdate` property is not needed to be public, this
privatizes it. It also removes the somewhat duplicated purpose of the
`clean_on_model_change` property, as we should rely on the solvers to
clean themselves up properly upon deletion instead of manually
triggering it.