Minor improvements to UpdateIRLS stopping criteria
Ditch the hard-coded 1e-12 value that gets added to the `f_old` to avoid
division by zero: assign `f_change` to `np.inf` instead.
Remove the absolute value computation while comparing `1 - chi_factor`
with the `misfit_tolerance`. This way we avoid ending the inversion with
a chi factor above the 1.0.