Implement magnetic G as linear operator (#1634)
Make the `G` property in the magnetic survey to be either a 2D Numpy
array, or a `scipy.sparse.LinearOperator` if `store_sensitivities` is
set to `"forward_only"`. The `LinearOperator` is capable of performing
operations like `G @ m` and `G.T @ v` without allocating the full `G`
matrix on memory or disk. This allows `Jvec` and `Jtvec` to be called
and not allocate the full `G` matrix. Overwrite the `getJ` method to
return either a 2D Numpy array or a `LinearOperator` formed by the
product between `G` and the `chiDeriv` as a `LinearOperator` as well.
Reimplement the `getJtJdiag` method: allow it to get the diagonal
without building `G` if `"forward_only"`, optimize the computation of
the diagonal when `G` is fully formed through `einsum`s, and fix the
cache of `gtg_diagonal` by checking the hash of the `W` argument. Raise
`NotImplementedError`s in some methods when `is_amplitude_data` is
`True`, specially the `"forward_only"` cases where we don't want to
allocate the `G` matrix. Extend docstrings of relevant methods. Add
tests for all the new features.