Clean up Numba functions for potential field simulations (#1663)
Declutter the potential field simulations by defining dictionaries with
Numba functions for gravity and magnetics. This way the simulations
don't need to define private attributes for each one of them: they can
just get the desired function in a single line. Raise
`NotImplementedError` on diagonal of G.T @ G when `"forward_only"` and
using `geoana` as engine. Specify which type of error should be raised
in the `xfail` parameters present in some tests.