Ensure misfit is purely real valued (#1524)
The misfit is always real valued, even for complex data, as it is the
dot-product of the complex-conjugate value with the value itself.
However, if the data is complex, `vdot(data, data)` will be a complex
number with zero complex part (`x+0j`). This ensures it gives back a
real-only value (`x`).