Use logging while setting default solver in PDE simulations
Ditch the `DefaultSolverWarning` in `get_default_solver` and make the
base PDE simulation class to handle the information message to users
about using a default solver. The information message is sent through
a custom logger defined using `logging`. Instead of being sent in the
`solver` setter (that would trigger repeated messages when working with
several simulations), the information is sent in the constructor if no
`solver` is passed. Add a new `verbose` flag to the base PDE class that
controls whether the info should be sent or not. Adjust tests
accordingly.