Allow use of `J` as `LinearOperator` in gravity equivalent layers (#1674)
Add Numba functions to compute `G.T @ v` and the diagonal of `G.T @ G`
for the gravity equivalent sources without storing the `G` matrix in
memory. Add needed private methods in the
`SimulationEquivalentSourceLayer`. Reorganize the private
`_numba_functions.py` file into a `_numba` submodule with files for the
functions related to the 3D and 2D meshes. Add tests for the new
functions.