Fix bug on `Impedance.eval` when orientation is "xx" or "yy" (#1692)
Make `Impedance.eval()` to return an array full of zeros in case that
the receiver's orientation is `"xx"` or "`yy`" and the mesh is not 3D.
This solves a bug that was causing the NSEM 1D simulation to error out
after trying to cast the output of `eval()` into an array. Add test for
the bugfix.

---------

Co-authored-by: Santiago Soler <santisoler@fastmail.com>