Update iteration print out (#1626)
Update the iteration write out order.

For the longest time, the iteration printers have written out mismatched
values at each iteration, and also did not print the last value of the
inversion.

This updates the write outs to happen after the line-search (matching
the behavior of the save directives), but before the directives
themselves alter the objective function.

I've also removed several un-needed notifications:
* setting bfgsH0
  * only set on BFGS subclasses and 
* if the UpdatePreconditioner directive is not in the directive list (It
handles the appoxHinv).
* Removes the repeated notifications about no nans in the model, and
instead raises an error if the model vector has any nans.

Writers:
* The iteration writer formatting logic was rather convoluted to me, so
I refactored it a bit to make it a little easier to understand what it
was doing...