Improve dipole source tests (#1711)
Improve the assertions in dipole source tests for EM. Compute the root
mean square of the difference between the numerical and analytic fields
in FDEM, and check it's below a certain tolerance. Use
`np.testing.assert_allclose` in TDEM tests to compare the predicted data
for both the B and H formulations.