Simple fix for pymatsolver 0.4.0  (#1717)
#### Fixes for pymatsolver 0.4.0
<!-- Add a summary of this Pull Request -->

#### PR Checklist
* [x] If this is a work in progress PR, set as a Draft PR
* [x] Linted my code according to the [style
guides](https://docs.simpeg.xyz/latest/content/getting_started/contributing/code-style.html).
* [x] Added
[tests](https://docs.simpeg.xyz/latest/content/getting_started/contributing/testing.html)
to verify changes to the code.
* [x] Added necessary documentation to any new functions/classes
following the
expect
[style](https://docs.simpeg.xyz/latest/content/getting_started/contributing/documentation.html).
* [x] Marked as ready for review (if this is was a draft PR), and
converted
      to a Pull Request
* [x] Tagged ``@simpeg/simpeg-developers`` when ready for review.

#### What does this implement/fix?
Squeeze these two locations in the SimulationDC operation (to be
consistent on output of `__inner_mat_mul `operation)

#### Additional information
The `__inner_mat_mul` function will "squeeze" out single length second
axes, so this fix causes it to be consistent for this operation in CG.
This function had to perform this way because of the way pymatsoler
previously handled single length array inputs.