Fix return of `get_indices_block` (#1713)
Make the `get_indices_block` function to always return an array of
integers for the cells that belong to the specified block, instead of
returning a tuple with just a single array. Improve its implementation.
Fix it for 1D mesh since it wasn't working properly. Improve error
messages: raise better errors than just `AssertError`s and improve
messages. Avoid overwriting the `p0` and `p1` arrays in-place if they
are not provided in the right order (west-south-bottom and then
east-north-top). Add tests for the function. Update usage of the
function across tests and examples. This introduces a backward
incompatible change.