Implement a closed loop as a TDEM source (#1651)
Implement the B field computation from a closed-loop wire source in the
TDEM code. We first compute the vector potential using the Biot-Savart
Law (implemented in geoana) and then take the curl to get the initial
B-field. Adds the computation of B with a step-off waveform using the B
or H formulations.

---------

Co-authored-by: Joseph Capriotti <josephrcapriotti@gmail.com>
Co-authored-by: Santiago Soler <santisoler@fastmail.com>