Relaxes LogisticSigmoidMap bound restrictions (#1741)
#### Summary
allows the sigmoid to map from `a` to `b` regardless of their respective
order, only errors if they are equal.

#### PR Checklist
* [x] If this is a work in progress PR, set as a Draft PR
* [x] Linted my code according to the [style
guides](https://docs.simpeg.xyz/latest/content/getting_started/contributing/code-style.html).
* [x] Added
[tests](https://docs.simpeg.xyz/latest/content/getting_started/contributing/testing.html)
to verify changes to the code.
* [x] Added necessary documentation to any new functions/classes
following the
expect
[style](https://docs.simpeg.xyz/latest/content/getting_started/contributing/documentation.html).
* [x] Marked as ready for review (if this is was a draft PR), and
converted
      to a Pull Request
* [x] Tagged ``@simpeg/simpeg-developers`` when ready for review.


#### What does this implement/fix?
Allows the sigmoid map to go arbitrarily between two numbers `a` and
`b`,

#### Additional information

having the sigmoid map the interval to [-5, 5] is just as valid as [5,
-5]